Generic types should work fine now across crates, in the view of rust/#9155 being fixed.
This also fixes one compiler warning.
