+++ src/components/main/layout/layout_task.rs
@@ -27,7 +27,8 @@ use geom::size::Size2D;
 use gfx::display_list::{ClipDisplayItemClass, DisplayItem, DisplayItemIterator, DisplayList};
 use gfx::font_context::{FontContext, FontContextInfo};
 use gfx::opts::Opts;
-use gfx::render_task::{RenderMsg, RenderChan, RenderLayer};
+use gfx::render_task::{RenderMsg, RenderLayer};
+use gfx::render_chan::{RenderChan};
 use gfx::{render_task, color};
 use script::dom::event::ReflowEvent;
 use script::dom::node::{ElementNodeTypeId, LayoutDataRef};
