+++ src/components/main/pipeline.rs
@@ -9,7 +9,8 @@ use extra::url::Url;
 use geom::size::Size2D;
 use gfx::opts::Opts;
 use gfx::render_task::{PaintPermissionGranted, PaintPermissionRevoked};
-use gfx::render_task::{RenderChan, RenderTask};
+use gfx::render_task::{RenderTask};
+use gfx::render_chan::{RenderChan};
 use layout::util::OpaqueNode;
 use script::layout_interface::LayoutChan;
 use script::script_task::LoadMsg;
